+++ docs/source/changelog.rst
@@ -21,6 +21,7 @@ Fixed
 * Pending interrupt messages are now removed from pika's queue before
   cancel is called.  This fixes an issue where an ``AtrributeError``
   was sometimes raised on worker shutdown.
+* Pika connection reset logs from the main thread are now hidden.
 
 
 `0.11.0`_ -- 2017-11-09