+++ docs/source/changelog.rst
@@ -6,8 +6,8 @@ Changelog
 All notable changes to this project will be documented in this file.
 
 
-`Unreleased`_
--------------
+`0.12.0`_ -- 2017-11-14
+-----------------------
 
 Added
 ^^^^^
@@ -201,6 +201,7 @@ Changed
 
 
 .. _Unreleased: https://github.com/Bogdanp/dramatiq/compare/v0.11.0...HEAD
+.. _0.12.0: https://github.com/Bogdanp/dramatiq/compare/v0.11.0...v0.12.0
 .. _0.11.0: https://github.com/Bogdanp/dramatiq/compare/v0.10.2...v0.11.0
 .. _0.10.2: https://github.com/Bogdanp/dramatiq/compare/v0.10.1...v0.10.2
 .. _0.10.1: https://github.com/Bogdanp/dramatiq/compare/v0.10.0...v0.10.1